Print List

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
 */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	if (h == NULL)
 	return (0);
@@ -26,7 +26,7 @@ size_t print_list(const list_t *h)
 			continue;
 		}
 
-		printf("[%d] %s\n", ptr->len, ptr->str);
+		printf("[%u] %s\n", ptr->len, ptr->str);
 		count++;
 		ptr = ptr->next;
